requirements.txt
fastapi
uvicorn
python-dotenv
pillow
pytesseract
gunicorn
azure-functions
requests
pydantic
anyio

